app-layer-htp: use file-store.stream-depth
Although file-store.stream-depth is enabled, a file
will be truncated if its size hits a limit (such
as request or response body) when a rule is used with
filestore keyword.

It happens because stream-depth is not evaluted, and as
a consequence files are truncated when they should not.
This patch adds a flag which is set when filestore is
used with http, and after that stream-depth set is now
considered.