eve: differentiate logging for alerted flows
This patch introduces a new eve-log option that add differentiate
logging for flows with alert. The concept is to get the maximal
verbosity for suspect flows so the maximum of context is provided
to the user. Alert contains metadata but for some alerts we don't
have the full transaction yet so we are missing information.

TLS and HTTP output are modified by this patch. For TLS, extended
logging is activated for flows with alerts. For HTTP, extended
logging is activated, full headers are dumped and HTTP body is
added.

Configuration allow to select if binary fields should be output
in base64 format or in printable format.