eve/http: use custom logging as option
This patch adds an option to select old logging style of HTTP
headers. The option is named `name-value-headers` and has to
be setup in http sections of eve. If set to `no` then logging
of headers is done in a Suricata 6.x style which is using a
object with key being the header name and value being the header
value.

Ticket: #6173