ci: add GPL mimetype data configure flag
As the mime type returned is dependant of the file system if we
don't have the flag, we will end up with tests depending of the
operating system files.

Also in operations such as build a cargo cache we need to have the
flag added to configure to have tree_magic_db accessible.